feat: add response handler util
