+++ product-service/src/constants/index.ts
