+++ product-service/src/data/products.ts
