+++ product-service/src/utils/index.ts
@@ -1,0 +1,7 @@
+import { commonHeaders } from '../constants';
+
+export const response = (statusCode, body, headers = commonHeaders) => ({
+  headers,
+  statusCode,
+  body: JSON.stringify(body),
+});
